mecab-api-server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mecab-python3==1.0.6
requests==2.31.0
youtube-transcript-api==1.1.0